requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.21.0
pillow>=9.0.0
tensorflow>=2.12.0
gdown>=4.6.0
pydantic>=2.0.0
